New SuffixNodeMemVec code, however needs some modification for SuffixNodeStore to start working again...
git-svn-id: svn+ssh://sgenomics.org/home/new/svn-repos/code/suffix4@331 dd0af01d-149e-4901-af2f-bc70ade6c2fb

+++ ChildStore.h
@@ -89,7 +89,7 @@ public:
   }
   
   int32_t get_last() {
-    for(size_t n=symbol_size-1;n>=0;n--) {
+    for(int32_t n=symbol_size-1;n>=0;n--) {
       if(m_symbols[n] >= 0) return m_symbols[n];
     }
     return -1;
